[FLINK-32420][refactor] Add the AbstractHeapPriorityQueueElement
